Initial commit
This is a resurrection of the "Norse"-project Harald worked on through
the final semester of "videregående" school. It has been updated to work
with the new way of storing xml-statistics through the StatsEditor, but
the stats are still not even close to being any good, they need a LOT of
tuning. The project has been renamed to fit the codename-convention of
norse creatures that Harald just randomly started doing.